// generated/src/app/api-models/procedure-code-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProcedureCodeMaster {
  procedureCode: string;
  shortDescription: string;
  asteriskProcedure: string;
  fullDescription: string;
  primaryGrouping: string;
  secondaryGrouping: string;
  followUpDays: number;
  codeClass: string;
  countAsDays: string;
  holdReason: string;
  holdDate: Date;
  bilateralAlternate: string;
  blockIndicator: string;
  blockStartProc: string;
  userDefined1: string;
  userDefined2: string;
  allowedReason: string;
  fromAge: number;
  toAge: number;
  patientGender: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  effectiveDate: Date;
  termDate: Date;
  targetRevCodeFlg: string;
  priceIndicatorFlg: string;
  productServiceIdQualifier: string;
  toothNumberRequired: string;
  surfaceRequired: string;
  archRequired: string;
  quadrantRequired: string;
  oralCavityRequired: string;
  weightedValue: number;
  userDate1: Date;
  userDate2: Date;
  instAuthDtlLnk: string;
}